+++ Assets/LDtkUnity/Editor/ParsedField/LDtkFieldParser.cs
@@ -3,11 +3,24 @@ using System.Linq;
 using UnityEngine;
 using UnityEngine.Internal;
 
-namespace LDtkUnity
+namespace LDtkUnity.Editor
 {
     [ExcludeFromDocs]
     public static class LDtkFieldParser
     {
+        private static LDtkBuilderEntity _builder;
+        
+        [RuntimeInitializeOnLoadMethod]
+        private static void ResetValue()
+        {
+            _builder = null;
+        }
+
+        public static void CacheRecentBuilder(LDtkBuilderEntity builder)
+        {
+            _builder = builder;
+        }
+        
         private static readonly List<ILDtkValueParser> ValueParsers = new List<ILDtkValueParser>
         {
             new LDtkParsedInt(),
@@ -26,8 +39,16 @@ namespace LDtkUnity
         public static ParseFieldValueAction GetParserMethodForType(FieldInstance instance)
         {
             ILDtkValueParser parser = ValueParsers.FirstOrDefault(p => p.TypeName(instance));
+            
+            
             if (parser != null)
             {
+                if (parser is ILDtkPostParser postParser)
+                {
+                    postParser.SupplyPostProcessorData(_builder, instance);
+                }
+                    
+                    
                 return parser.ImportString;
             }