+++ Assets/LDtkUnity/Editor/ParsedField/Parsed/ILDtkPostParser.cs
@@ -1,0 +1,7 @@
+﻿namespace LDtkUnity.Editor
+{
+    public interface ILDtkPostParser
+    {
+        void SupplyPostProcessorData(LDtkBuilderEntity builder, FieldInstance field);
+    }
+}