+++ Assets/LDtkUnity/Editor/ParsedField/Process/ILDtkPostParseProcess.cs
@@ -1,0 +1,7 @@
+﻿namespace LDtkUnity.Editor
+{
+    public interface ILDtkPostParseProcess<T>
+    {
+        T Postprocess(T value);
+    }
+}