+++ Assets/LDtkUnity/Runtime/Fields/PointParseData.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace LDtkUnity.Editor
+{
+    public struct PointParseData
+    {
+        public Vector2 LevelPosition;
+        public int LvlCellHeight;
+        public int PixelsPerUnit;
+        public int GridSize;
+    }
+}